+++ src/drivers/terminal.cpp
@@ -9,7 +9,7 @@ unsigned int StrLen(char* str)
     return i;
 }
 
-int IToA(int value, char *sp, int radix)
+int IToA(long value, char *sp, int radix)
 {
     char tmp[16];
     char *tp = tmp;
@@ -119,7 +119,7 @@ namespace Terminal
     }
 
 
-    Status PrintInt(unsigned int value, int radix, Graphics::Color color)
+    Status PrintInt(unsigned long value, int radix, Graphics::Color color)
     {
         char buffer[16];
         IToA(value, buffer, radix);
